Fix: Move war archive events where others are, adjust code as needed
