+++ campaign/war_archives_20190329_cn/b1.py
@@ -1,4 +1,4 @@
-from ..campaign_base import CampaignBase
+from ..campaign_war_archives.campaign_base import CampaignBase
 from module.map.map_base import CampaignMap
 from module.map.map_grids import SelectedGrids, RoadGrids
 from module.logger import logger
@@ -17,9 +17,9 @@ MAP.map_data = """
 MAP.weight_data = """
     50 50 10 50 30 20 10 10 10 50
     50 50 50 50 30 50 20 50 50 10
-    50 50 50 50 50 50 50 50 30 50
+    50 50 50 50 50 50 50 50 10 50
     50 50 50 50 50 50 20 50 50 10
-    50 50 50 10 30 20 10 10 10 50
+    50 50 50 10 20 20 10 10 10 50
 """
 MAP.spawn_data = [
     {'battle': 0, 'enemy': 4},
