+++ campaign/war_archives_20190329_cn/b3.py
@@ -1,4 +1,4 @@
-from ..campaign_base import CampaignBase
+from ..campaign_war_archives.campaign_base import CampaignBase
 from module.map.map_base import CampaignMap
 from module.map.map_grids import SelectedGrids, RoadGrids
 from module.logger import logger
@@ -22,13 +22,13 @@ MAP.map_data = """
 """
 MAP.weight_data = """
     50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
-    50 50 50 50 50 50
+    50 10 50 50 10 50
+    50 10 50 50 10 50
+    50 10 40 40 10 50
+    50 20 40 40 20 50
+    50 30 20 20 30 50
+    20 50 50 50 50 20
+    20 20 50 50 20 20
     50 50 50 50 50 50
     50 50 50 50 50 50
 """
