+++ campaign/war_archives_20190329_cn/c2.py
@@ -1,34 +1,32 @@
-from ..campaign_base import CampaignBase
+from ..campaign_war_archives.campaign_base import CampaignBase
 from module.map.map_base import CampaignMap
 from module.map.map_grids import SelectedGrids, RoadGrids
 from module.logger import logger
-from .a1 import Config as ConfigBase
+from .c1 import Config as ConfigBase
 
-MAP = CampaignMap('A3')
-MAP.shape = 'H7'
-MAP.camera_data = ['D2', 'D5', 'E2', 'E5']
-MAP.camera_data_spawn_point = ['D2']
+MAP = CampaignMap('C2')
+MAP.shape = 'H6'
+MAP.camera_data = ['D2', 'D4', 'E2', 'E4']
+MAP.camera_data_spawn_point = ['D4']
 MAP.map_data = """
-    SP -- -- ++ ++ ++ ++ ++
-    -- ++ ME -- ME -- -- ++
-    SP -- -- ME ME -- ME ++
-    ++ ++ ++ ++ ++ ME -- ++
-    ++ MB ME ++ ++ -- ME --
-    ++ ME ME ME ME ME ++ --
-    ++ ME -- ME ME -- ME --
+    -- -- ++ ++ ++ ++ -- MM
+    ME -- ME -- ME -- ME --
+    -- ++ -- ++ ++ ME ++ ME
+    -- -- ME ++ ++ ME ++ --
+    -- -- ++ MB ME ME -- --
+    SP SP ++ MB ME ME ME --
 """
 MAP.weight_data = """
     50 50 50 50 50 50 50 50
-    50 50 20 50 20 50 50 50
-    50 50 50 30 20 50 50 50
+    20 50 10 10 10 10 50 50
     50 50 50 50 50 10 50 50
-    50 10 10 50 50 10 50 50
-    50 10 10 10 10 10 50 50
-    50 50 50 50 50 50 50 50
+    50 50 20 50 50 10 50 50
+    50 50 50 10 10 10 50 50
+    50 50 50 10 20 20 50 50
 """
 MAP.spawn_data = [
     {'battle': 0, 'enemy': 4},
-    {'battle': 1, 'enemy': 1},
+    {'battle': 1, 'enemy': 2, 'mystery': 1},
     {'battle': 2, 'enemy': 1},
     {'battle': 3, 'enemy': 1},
     {'battle': 4, 'enemy': 1},
@@ -41,7 +39,6 @@ A3, B3, C3, D3, E3, F3, G3, H3, \
 A4, B4, C4, D4, E4, F4, G4, H4, \
 A5, B5, C5, D5, E5, F5, G5, H5, \
 A6, B6, C6, D6, E6, F6, G6, H6, \
-A7, B7, C7, D7, E7, F7, G7, H7, \
     = MAP.flatten()
