+++ campaign/war_archives_20190329_cn/c3.py
@@ -1,24 +1,29 @@
-from ..campaign_base import CampaignBase
+from ..campaign_war_archives.campaign_base import CampaignBase
 from module.map.map_base import CampaignMap
 from module.map.map_grids import SelectedGrids, RoadGrids
 from module.logger import logger
+from .c1 import Config as ConfigBase
 
-MAP = CampaignMap('A1')
-MAP.shape = 'H5'
-MAP.camera_data = ['D2', 'D3', 'E2', 'E3']
-MAP.camera_data_spawn_point = ['D3', 'D2']
+MAP = CampaignMap('C3')
+MAP.shape = 'H7'
+MAP.camera_data = ['D2', 'D5', 'E2', 'E5']
+MAP.camera_data_spawn_point = ['D2']
 MAP.map_data = """
     SP -- -- ++ ++ ++ ++ ++
-    -- ++ ME ME ME -- ME ++
-    SP -- ME -- ME ++ ME MB
-    -- -- ++ ME ME ++ ME MB
-    SP -- ++ -- ME -- ME MB
+    -- ++ ME -- ME -- -- ++
+    SP -- -- ME ME -- ME ++
+    ++ ++ ++ ++ ++ ME -- ++
+    ++ MB ME ++ ++ -- ME --
+    ++ ME ME ME ME ME ++ --
+    ++ ME -- ME ME -- ME --
 """
 MAP.weight_data = """
     50 50 50 50 50 50 50 50
-    50 50 50 50 50 50 50 50
-    50 50 50 50 50 50 50 50
-    50 50 50 50 50 50 50 50
+    50 50 20 50 20 50 50 50
+    50 50 50 30 20 50 50 50
+    50 50 50 50 50 10 50 50
+    50 10 10 50 50 10 50 50
+    50 10 10 10 10 10 50 50
     50 50 50 50 50 50 50 50
 """
 MAP.spawn_data = [
@@ -27,17 +32,20 @@ MAP.spawn_data = [
     {'battle': 2, 'enemy': 1},
     {'battle': 3, 'enemy': 1},
     {'battle': 4, 'enemy': 1},
-    {'battle': 5, 'boss': 1},
+    {'battle': 5, 'enemy': 1},
+    {'battle': 6, 'boss': 1},
 ]
 A1, B1, C1, D1, E1, F1, G1, H1, \
 A2, B2, C2, D2, E2, F2, G2, H2, \
 A3, B3, C3, D3, E3, F3, G3, H3, \
 A4, B4, C4, D4, E4, F4, G4, H4, \
 A5, B5, C5, D5, E5, F5, G5, H5, \
+A6, B6, C6, D6, E6, F6, G6, H6, \
+A7, B7, C7, D7, E7, F7, G7, H7, \
     = MAP.flatten()
 
 
-class Config:
+class Config(ConfigBase):
     # ===== Start of generated config =====
     MAP_HAS_MAP_STORY = True
     MAP_HAS_FLEET_STEP = False
@@ -49,7 +57,10 @@ class Campaign(CampaignBase):
     MAP = MAP
 
     def battle_0(self):
+        if self.fleet_2_push_forward():
+            return True
+
         return self.battle_default()
 
-    def battle_5(self):
+    def battle_6(self):
         return self.fleet_boss.clear_boss()
