+++ config/template.ini
@@ -174,8 +174,8 @@ sos_fleets_chapter_9 = 5, 6, 1
 sos_fleets_chapter_10 = 5, 6, 1
 
 [War_archives]
-war_archives_stage = d3
-war_archives_name = event_20200326_cn
+war_archives_stage = a1
+war_archives_name = war_archives_20190329_cn
 
 [Raid]
 raid_name = raid_20200624
