+++ module/campaign/run.py
@@ -40,7 +40,7 @@ class CampaignRun(Reward):
 
         if folder.startswith('campaign_'):
             self.stage = '-'.join(name.split('_')[1:3])
-        if folder.startswith('event'):
+        if folder.startswith('event') or folder.startswith('war_archives'):
             self.stage = name
 
         try:
